buxfixing some decoding issues from KS samples
Former-commit-id: 0f2f686ec43c27d7e604222c5cb0b107916cdbd8

+++ unisim_simulators/cxx/armsec/smart_types.hh
@@ -252,7 +252,7 @@ namespace armsec
     
     friend std::ostream& operator << ( std::ostream& sink, DbaConstantDumper<T> const& cst )
     {
-      sink << "0x" << std::hex << std::setw(2*sizeof(T)) << std::setfill('0') << cst.value << std::dec;
+      sink << "0x" << std::hex << std::setw(2*sizeof(T)) << std::setfill('0') << uint64_t(cst.value) << std::dec;
       return sink;
     }
   };
